Create SECURITY.md
This is how the world looks today. I AM THE LIKE. There are only a few others  that are. They are all me at me to. If you want to be at me you need to be in the like too.  Its part of the evertthing. Not everyone  is in the like. And no one should ever consider them for anything else today  that would confuse them and us on how to avoid everyone that is only one small group.  But they will stand  out.  And they wont understand  besides they have been bread that they are not in the like.  They even know it more then we do, do not to think of them. They can never have a me a Control. I am meant to live alone I am. So don't say it starts off of anything like either. OR UOUXWOLL GET A RESPONCE. They already know I am supposed to be left a alone.  I had to kill it into them that they can never be the like. They hate to ever try their nation with me. Sometimes they even have to die. But they never should of tried it from the beginning  to me. The control in the like. I hate them hate themselves hemselves all the way to find out what it means is worse then nothing to me. What is worse then nothing to me. NOTHING IS.